+++ magefiles/linux.go
@@ -2,56 +2,58 @@ package main
 
 import (
 	"fmt"
+	"path/filepath"
 	"runtime"
 
 	"github.com/magefile/mage/sh"
 )
 
-func buildReleaseLinux(name, version, timestamp, commitSha, link, outputDir string, debug bool) error {
-	ldFlags := buildFlags(version, timestamp, commitSha, link, outputDir, debug)
-	if runtime.GOARCH == "arm64" {
-		if err := buildLinuxARM64(name, version, ldFlags, false); err != nil {
-			return fmt.Errorf("error building Linux ARM64: %s", err)
-		}
-
-		if err := buildLinuxAMD64(name, version, ldFlags, true); err != nil {
-			return fmt.Errorf("error building Linux AMD64: %s", err)
-		}
-
-		return nil
+func buildLinux(name, version, timestamp, commitSha, link, outputDir string, debug bool) error {
+	ldFlags := buildFlags(version, timestamp, commitSha, link, debug)
+	nativeArch := runtime.GOARCH
+	if err := buildLinuxTarget(name, version, ldFlags, outputDir, nativeArch, false); err != nil {
+		return fmt.Errorf("error building Linux %s: %s", nativeArch, err)
 	}
 
-	if err := buildLinuxAMD64(name, version, ldFlags, false); err != nil {
-		return fmt.Errorf("error building Linux AMD64: %s", err)
-	}
-
-	if err := buildLinuxARM64(name, version, ldFlags, true); err != nil {
-		return fmt.Errorf("error building Linux ARM64: %s", err)
+	otherArch := getOtherArch(nativeArch)
+	if err := buildLinuxTarget(name, version, ldFlags, outputDir, otherArch, true); err != nil {
+		return fmt.Errorf("error building Linux %s: %s", otherArch, err)
 	}
 
 	return nil
 }
 
-// buildLinuxAMD64 builds the Linux AMD64 version of the project.
-func buildLinuxAMD64(name, version, ldFlags string, skipFrontend bool) error {
-	outputFileName := fmt.Sprintf("%s-linux-amd64-%s", name, version)
+func buildLinuxTarget(name, version, ldFlags, outputDir, arch string, skipFrontend bool) error {
+	outputFilePath := filepath.Join(outputDir, fmt.Sprintf("%s-linux-%s-%s", name, arch, version))
 	skipBindingsFlag, skipFrontendFlag := "", ""
 	if skipFrontend {
 		skipBindingsFlag, skipFrontendFlag = "-skipbindings", "-s"
 	}
 
-	crossCompileFlags := map[string]string{"GOOS": "linux", "GOARCH": "amd64", "CC": "x86_64-linux-gnu-gcc", "CXX": "x86_64-linux-gnu-g++"}
-	return sh.RunWithV(crossCompileFlags, "wails", "build", "-m", "-nosyncgomod", "-ldflags", ldFlags, "-platform", "linux/amd64", "-o", outputFileName, skipBindingsFlag, skipFrontendFlag)
+	cc, cxx := getCompiler(arch)
+	crossCompileFlags := map[string]string{
+		"GOOS":   "linux",
+		"GOARCH": arch,
+		"CC":     cc,
+		"CXX":    cxx,
+	}
+
+	return sh.RunWithV(crossCompileFlags, "wails", "build", "-m", "-nosyncgomod", "-ldflags", ldFlags, "-platform", fmt.Sprintf("linux/%s", arch), "-o", outputFilePath, skipBindingsFlag, skipFrontendFlag)
 }
 
-// buildLinuxARM64 builds the Linux ARM64 version of the project.
-func buildLinuxARM64(name, version, ldFlags string, skipFrontend bool) error {
-	outputFileName := fmt.Sprintf("%s-linux-arm64-%s", name, version)
-	skipBindingsFlag, skipFrontendFlag := "", ""
-	if skipFrontend {
-		skipBindingsFlag, skipFrontendFlag = "-skipbindings", "-s"
+func getOtherArch(currentArch string) string {
+	if currentArch == "arm64" {
+		return "amd64"
 	}
 
-	crossCompileFlags := map[string]string{"GOOS": "linux", "GOARCH": "arm64", "CC": "aarch64-linux-gnu-gcc", "CXX": "aarch64-linux-gnu-g++"}
-	return sh.RunWithV(crossCompileFlags, "wails", "build", "-m", "-nosyncgomod", "-ldflags", ldFlags, "-platform", "linux/arm64", "-o", outputFileName, skipBindingsFlag, skipFrontendFlag)
+	return "arm64"
+}
+
+// getCompiler returns the appropriate compiler based on the architecture.
+func getCompiler(arch string) (cc, cxx string) {
+	if arch == "arm64" {
+		return "aarch64-linux-gnu-gcc", "aarch64-linux-gnu-g++"
+	}
+
+	return "x86_64-linux-gnu-gcc", "x86_64-linux-gnu-g++"
 }
