+++ magefiles/windows.go
@@ -2,13 +2,14 @@ package main
 
 import (
 	"fmt"
+	"path/filepath"
 
 	"github.com/magefile/mage/sh"
 )
 
-func buildReleaseWindows(name, version, timestamp, commitSha, link, outputDir string, debug bool) error {
-	ldFlags := buildFlags(version, timestamp, commitSha, link, outputDir, debug)
-	if err := buildWindowsAMD64(name, version, ldFlags, true); err != nil {
+func buildWindows(name, version, timestamp, commitSha, link, outputDir string, debug bool) error {
+	ldFlags := buildFlags(version, timestamp, commitSha, link, debug)
+	if err := buildWindowsAMD64(name, version, ldFlags, outputDir, true); err != nil {
 		return fmt.Errorf("error building Windows AMD64: %s", err)
 	}
 
@@ -16,8 +17,8 @@ func buildReleaseWindows(name, version, timestamp, commitSha, link, outputDir st
 }
 
 // buildWindowsAMD64 builds the Windows AMD64 version of the project.
-func buildWindowsAMD64(name, version, ldFlags string, skipFrontend bool) error {
-	outputFileName := fmt.Sprintf("%s-windows-amd64-%s.exe", name, version)
+func buildWindowsAMD64(name, version, ldFlags, outputDir string, skipFrontend bool) error {
+	outputFilePath := filepath.Join(outputDir, fmt.Sprintf("%s-windows-amd64-%s.exe", name, version))
 	skipBindingsFlag, skipFrontendFlag := "", ""
 	if skipFrontend {
 		skipBindingsFlag, skipFrontendFlag = "-skipbindings", "-s"
@@ -30,7 +31,7 @@ func buildWindowsAMD64(name, version, ldFlags string, skipFrontend bool) error {
 		"CXX":    "x86_64-w64-mingw32-g++",
 	}
 
-	err := sh.RunWithV(crossCompileFlags, "wails", "build", "-m", "-nosyncgomod", "-ldflags", ldFlags, "-nsis", "-platform", "windows/amd64", "-o", outputFileName, skipBindingsFlag, skipFrontendFlag)
+	err := sh.RunWithV(crossCompileFlags, "wails", "build", "-m", "-nosyncgomod", "-ldflags", ldFlags, "-nsis", "-platform", "windows/amd64", "-o", outputFilePath, skipBindingsFlag, skipFrontendFlag)
 	if err != nil {
 		return fmt.Errorf("error building Windows AMD64: %v", err)
 	}
